+++ examples/P1P2Monitor/P1P2_Daikin_ParameterConversion.h
@@ -1,5 +1,5 @@
 /* PP1P2_Daikin_ParameterConversion.h product-ndependent code
- *                     
+ *
  * Copyright (c) 2019 Arnold Niessen, arnold.niessen -at- gmail-dot-com  - licensed under GPL v2.0 (see LICENSE)
  *
  */
@@ -33,19 +33,23 @@ float FN_f8s8(uint8_t *b)            { return ((float)(int8_t) b[-1]) + ((float)
 
 #ifdef SAVEHISTORY
 // save-history pointers
-static byte rbhistory[RBHLEN];           // history storage
-static uint16_t savehistoryend=0;        // #bytes saved so far
-static uint16_t savehistoryp[RBHNP]={};  // history pointers
-static  uint8_t savehistorylen[RBHNP]={}; // length of history for a certain packet
+static byte rbhistory[RBHLEN];              // history storage
+static uint16_t savehistoryend=0;           // #bytes saved so far
+static uint16_t savehistoryp[RBHNP] = {};   // history pointers
+static  uint8_t savehistorylen[RBHNP] = {}; // length of history for a certain packet
+static float savehistoryvalue[NHYST] = {};
+static float saveavgvalue[NHYST] = {};
+static bool valsaved[NHYST] = {};
 
-static byte paramhistory[PARAMLEN * 4];      // history storage for params in 00F035/00F135/40F035/40F13
+static byte param35history[PARAM35LEN * 4];     // history storage for params in 00F035/00F135/40F035/40F135
+static uint16_t param36history[PARAM36LEN * 4]; // history storage for params in 00F036/00F136/40F036/40F136
 
-int8_t savehistoryindex(byte *rb); // defined in product-dependent code
+int8_t savehistoryindex(byte *rb);          // defined in product-dependent code
 uint8_t savehistoryignore(byte *rb);
 
 void savehistory(byte *rb, int n) {
 // also modifies savehistoryp, savehistorylen, savehistoryend
-  if (n > 3) {  
+  if (n > 3) {
     int8_t shi = savehistoryindex(rb);
     if (shi >= 0) {
       uint8_t shign = savehistoryignore(rb);
@@ -116,13 +120,43 @@ void savehistory(byte *rb, int n) {
     } else if (shi == -0x35) {
       // save params in Fx35 packets
       for (byte i = 5; i < n; i += 3) {
-        int paramnr = (((uint16_t) rb[i-1]) << 8) | rb[i-2];
+        uint16_t paramnr = (((uint16_t) rb[i-1]) << 8) | rb[i-2];
         byte paramcat = ((rb[0] & 0x40) >> 5) + (rb[1] & 0x01);
         byte paramval = rb[i];
-        if ((paramnr >= PARAMSTART) && (paramnr < PARAMSTART + PARAMLEN)) paramhistory[paramnr - PARAMSTART + PARAMLEN * paramcat] = paramval;
+        if ((paramnr >= PARAM35START) && (paramnr < PARAM35START + PARAM35LEN)) param35history[paramnr - PARAM35START + PARAM35LEN * paramcat] = paramval;
+      }
+    } else if (shi == -0x36) {
+      // save params in Fx36 packets
+      for (byte i = 6; i < n; i += 4) {
+        uint16_t paramnr = (((uint16_t) rb[i-2]) << 8) | rb[i-3];
+        byte paramcat = ((rb[0] & 0x40) >> 5) + (rb[1] & 0x01);
+        uint16_t paramval = (((uint16_t) rb[i]) << 8) | rb[i-1];
+        if ((paramnr >= PARAM36START) && (paramnr < PARAM36START + PARAM36LEN)) param36history[paramnr - PARAM36START + PARAM36LEN * paramcat] = paramval;
       }
     }
   }
+}
+
+bool newmeasuredval(float f, byte i, float hyst1, float hyst2) {
+// hyst1 should be (say 20%) more than step-size of measured parameter
+// hyst2 should be somewhat less (say 80%) than step-size of measured parameter
+//    for 0.5 degree resolution, we use hyst1=0.4 hyst2=0.6
+//    for 0.03 degree resolution with lots of variation, we use hyst1=0.09 hyst2=0.15
+  if (i >= NHYST) return 1; // no history, assume value is new
+  if (!valsaved[i]) { savehistoryvalue[i] = saveavgvalue[i] = f; valsaved[i] = 1; return 1;} // first time we see this parameter
+  float avg = 0.9 * saveavgvalue[i] + 0.1 * f;
+  saveavgvalue[i] = avg;
+  // check if value changed more than hyst2 (high threshold) compared to historic (non-averaged) value
+  if ((f > savehistoryvalue[i] + hyst2) || (f < savehistoryvalue[i] - hyst2)) {
+    savehistoryvalue[i] = f;
+    return 1;
+  }
+  // check if averaged value changed more than hyst1 (low threshold) compared to historic (non-averaged) value
+  if ((avg > savehistoryvalue[i] + hyst1) || (avg < savehistoryvalue[i] - hyst1)) {
+    savehistoryvalue[i] = f;
+    return 1;
+  }
+  return !changeonly;
 }
 
 bool newbytesval(byte *rb, byte i, byte k) {
@@ -148,8 +182,20 @@ bool newbytesval(byte *rb, byte i, byte k) {
     int paramnr = (((uint16_t) rb[i-1]) << 8) | rb[i-2];
     byte paramcat = ((rb[0] & 0x40) >> 5) + (rb[1] & 0x01);
     byte paramval = rb[i];
-    if ((paramnr >= PARAMSTART) && (paramnr < PARAMSTART + PARAMLEN)) {
-      if (paramhistory[paramnr - PARAMSTART + PARAMLEN * paramcat] != paramval) return 1; else return !changeonly;
+    if ((paramnr >= PARAM35START) && (paramnr < PARAM35START + PARAM35LEN)) {
+      if (param35history[paramnr - PARAM35START + PARAM35LEN * paramcat] != paramval) return 1; else return !changeonly;
+    } else {
+      // paramval outside saved range
+      return 1;
+    }
+  } else if (shi == -0x36) {
+    // check param in Fx36 packet
+    // check whether param nr (i-3)/(i-2) in byte (i-1)/i has a new value (also returns true if byte has not been saved or if changeonly=0)
+    uint16_t paramnr = (((uint16_t) rb[i-2]) << 8) | rb[i-3];
+    byte paramcat = ((rb[0] & 0x40) >> 5) + (rb[1] & 0x01);
+    uint16_t paramval = (((uint16_t) rb[i]) << 8) | rb[i-1];
+    if ((paramnr >= PARAM36START) && (paramnr < PARAM36START + PARAM36LEN)) {
+      if (param36history[paramnr - PARAM36START + PARAM36LEN * paramcat] != paramval) return 1; else return !changeonly;
     } else {
       // paramval outside saved range
       return 1;
@@ -185,17 +231,30 @@ bool newbitval(byte *rb, byte i, byte p) {
   }
 }
 #else /* !SAVEHISTORY */
+#define newmeasuredval(a, b, c, d) (1)
 #define newbytesval(a, b, c) (1)
 #define newbitval(a, b, c) (1)
 #endif /* SAVEHISTORY */
 
 byte handleparam(char* key, char* value, byte* rb, uint8_t i) {
-  int paramnr = (((uint16_t) rb[i-1]) << 8) | rb[i-2];
-  byte paramval = rb[i];
+  uint16_t paramnr = 0;
+  uint16_t paramval = 0;
+  if (rb[2] == 0x35) {
+    paramnr = (((uint16_t) rb[i-1]) << 8) | rb[i-2];
+    paramval = rb[i];
+  } else if (rb[2] == 0x36) {
+    paramnr = (((uint16_t) rb[i-2]) << 8) | rb[i-3];
+    paramval = (((uint16_t) rb[i]) << 8) | rb[i-1];
+  }
   //TODO if (!outputunknown) return 0;
   if (!newbytesval(rb, i, 0)) return 0;
   if (paramnr == 0xFFFF) return 0; // no param
-  if (paramnr == 0xA2) return 0; // counter; useless?
+  if (paramnr == 0xA2) return 0; // counter; useless?; TODO  is this device-independent?
+
+  // TODO get key for known params
+  // byte rv = paramnr2key(rb, paramnr, paramval, j, key, value);
+  // try to avoid repetition of code execution in bitbasis mode
+
   snprintf(key, KEYLEN, "P0x%X%X-0x%X%X-0x%X%X-0x%X%X", rb[0] >> 4, rb[0]&0x0F, rb[1] >> 4, rb[1] & 0x0F, rb[2] >> 4, rb[2] & 0x0F, paramnr >> 4, paramnr & 0x0F);
   snprintf(value, KEYLEN, "0x%2X", paramval);
   return 1;
@@ -219,29 +278,64 @@ byte unknownbit(char* key, char* value, byte* rb, uint8_t i, uint8_t j) {
 #define UNKNOWN     {(cat = 0);}
 #define SETTING     {(cat = 1);}
 #define MEASUREMENT {(cat = 2);}
-#define PARAM35     {(cat = 3);}
+#define PARAM3x     {(cat = 3);}
 #define TEMPFLOWP   {(cat = 4);}
 
 #define BITBASIS         { return newbytesval(rb, i, 1) << 3; } // returns 8 if at least one bit of a byte changed, otherwise 0
-#define VALUE_u8         { if (!newbytesval(rb, i, 1)) return 0; snprintf(value, KEYLEN, "%i", rb[i]);                return 1; }
-#define VALUE_u24        { if (!newbytesval(rb, i, 3)) return 0; snprintf(value, KEYLEN, "%lui", FN_u24(&rb[i]));     return 1; }
+
+// Some #defines are replaced to reduce program memory size by approximately 5 kB
+// Other #defines are not replaced as it does not save any program memory - perhaps compiler optimization related?
+//
+uint8_t value_u8(byte* rb, uint8_t i, char* value) {
+  if (!newbytesval(rb, i, 1)) return 0;
+  snprintf(value, KEYLEN, "%i", rb[i]);
+  return 1;
+}
+
+//#define VALUE_u8         { if (!newbytesval(rb, i, 1)) return 0; snprintf(value, KEYLEN, "%i", rb[i]);                return 1; }
+#define VALUE_u8         return value_u8(rb, i, value);
+
+// function instead of macro, saves code size
+uint8_t value_u24(byte* rb, uint8_t i, char* value) {
+  if (!newbytesval(rb, i, 3)) return 0;
+  snprintf(value, KEYLEN, "%li", FN_u24(&rb[i]));
+  return 1;
+}
+
+/* function instead of macro, but does not seem to save code size
+uint8_t value_f8_8(byte* rb, uint8_t i, char* value, uint8_t cnt, float hyst1, float hyst2) {
+  float f= FN_f8_8(&rb[i]);
+  if (!newmeasuredval(f, cnt, hyst1, hyst2)) return 0;
+  dtostrf(f, 1, 2, value);
+  return 1;
+}*/
+
+// function instead of macro, saves code size
+uint8_t value_flag8(byte* rb, uint8_t i, uint8_t j, char* value) {
+  if (!newbitval(rb, i, j))   return 0;
+  snprintf(value, KEYLEN, "%i", FN_flag8(rb[i], j));
+  return 1;
+}
+
+#define VALUE_u8hex      { if (!newbytesval(rb, i, 1)) return 0; snprintf(value, KEYLEN, "0x%x", rb[i]);              return 1; }
+#define VALUE_u24        return value_u24(rb, i, value);
+#define VALUE_flag8      return value_flag8(rb, i, j, value);
 #define VALUE_u8_add2k   { if (!newbytesval(rb, i, 1)) return 0; snprintf(value, KEYLEN, "%i", rb[i]+2000);           return 1; }
 #define VALUE_u8delta    { if (!newbytesval(rb, i, 1)) return 0; snprintf(value, KEYLEN, "%i", FN_u8delta(&rb[i]));   return 1; }
-#define VALUE_flag8      { if (!newbitval(rb, i, j))   return 0; snprintf(value, KEYLEN, "%i", FN_flag8(rb[i], j));   return 1; }
 #ifdef RPI
-#define VALUE_f8_8       { if (!newbytesval(rb, i, 2)) return 0; snprintf(value, KEYLEN, "%11f", FN_f8_8(&rb[i]));    return 1; }
-#define VALUE_f8s8       { if (!newbytesval(rb, i, 2)) return 0; snprintf(value, KEYLEN, "%11f", FN_f8s8(&rb[i]));    return 1; }
-#define VALUE_u8div10    { if (!newbytesval(rb, i, 1)) return 0; snprintf(value, KEYLEN, "%11f", FN_u8div10(&rb[i])); return 1; }
-#define VALUE_F(v, ch)   { if (changeonly && !ch)      return 0; snprintf(value, KEYLEN, "%11f", v);                  return 1; }
+#define VALUE_f8_8(cnt, hyst1, hyst2)    { if (!newmeasuredval(FN_f8_8(&rb[i]),    cnt, hyst1, hyst2)) return 0; snprintf(value, KEYLEN, "%1.1f", FN_f8_8(&rb[i]));    return 1; }
+#define VALUE_f8s8(cnt, hyst1, hyst2)    { if (!newmeasuredval(FN_f8s8(&rb[i]),    cnt, hyst1, hyst2)) return 0; snprintf(value, KEYLEN, "%1.1f", FN_f8s8(&rb[i]));    return 1; }
+#define VALUE_u8div10(cnt, hyst1, hyst2) { if (!newmeasuredval(FN_u8div10(&rb[i]), cnt, hyst1, hyst2)) return 0; snprintf(value, KEYLEN, "%1.1f", FN_u8div10(&rb[i])); return 1; }
+#define VALUE_F(v, cnt, hyst1, hyst2)    { if (changeonly && !newmeasuredval(v,    cnt, hyst1, hyst2)) return 0; snprintf(value, KEYLEN, "%1.1f", v);                  return 1; }
 #else /* RPI */
 // note that snprintf(.. , "%f", ..) is not supported on Arduino/Atmega so use dtostrf instead
-#define VALUE_f8_8       { if (!newbytesval(rb, i, 2)) return 0; dtostrf(FN_f8_8(&rb[i]), 1, 2, value);               return 1; }
-#define VALUE_f8s8       { if (!newbytesval(rb, i, 2)) return 0; dtostrf(FN_f8s8(&rb[i]), 1, 1, value);               return 1; }
-#define VALUE_u8div10    { if (!newbytesval(rb, i, 1)) return 0; dtostrf(FN_u8div10(&rb[i]), 1, 1, value);            return 1; }
-#define VALUE_F(v, ch)   { if (changeonly && !ch)      return 0; dtostrf(v, 1, 1, value);                             return 1; }
+#define VALUE_f8_8(cnt, hyst1, hyst2)    { if (!newmeasuredval(FN_f8_8(&rb[i]),    cnt, hyst1, hyst2)) return 0; dtostrf(FN_f8_8(&rb[i]), 1, 2, value);               return 1; }
+#define VALUE_f8s8(cnt, hyst1, hyst2)    { if (!newmeasuredval(FN_f8s8(&rb[i]),    cnt, hyst1, hyst2)) return 0; dtostrf(FN_f8s8(&rb[i]), 1, 1, value);               return 1; }
+#define VALUE_u8div10(cnt, hyst1, hyst2) { if (!newmeasuredval(FN_u8div10(&rb[i]), cnt, hyst1, hyst2)) return 0; dtostrf(FN_u8div10(&rb[i]), 1, 1, value);            return 1; }
+#define VALUE_F(v, cnt, hyst1, hyst2)    { if (changeonly && !newmeasuredval(v,    cnt, hyst1, hyst2)) return 0; dtostrf(v, 1, 1, value);                             return 1; }
 #endif /* RPI */
 #define VALUE_H4         { if (!newbytesval(rb, i, 4)) return 0; snprintf(value, KEYLEN, "%X%X%X%X%X%X%X%X", rb[i-3] >> 4, rb[i-3] & 0x0F, rb[i-2] >> 4, rb[i-2] & 0x0F, rb[i-1] >> 4, rb[i-1] & 0x0F, rb[i] >> 4, rb[i] & 0x0F); return 1; }
-#define HANDLEPARAM      { PARAM35; return handleparam(key, value, rb, i);}
+#define HANDLEPARAM      { PARAM3x; return handleparam(key, value, rb, i);}
 #define UNKNOWNBYTE      { UNKNOWN; return unknownbyte(key, value, rb, i);}
 #define UNKNOWNBIT       { UNKNOWN; return unknownbit(key, value, rb, i, j);}
 #define BITBASIS_UNKNOWN { switch (j) { case 8 : BITBASIS; default : UNKNOWNBIT; } }